cleaning(core): move manifest to node-utils
