+++ packages/core/src/filesystem/index.ts
@@ -1,4 +1,3 @@
-export * from "./manifest";
 export * from "./paths";
 export * from "./pkg";
 export * from "./PrismicSharedConfig";
