+++ packages/core/src/node-utils/index.ts
@@ -1,1 +1,2 @@
 export * as Framework from "./framework";
+export * from "./manifest";
