+++ packages/slice-machine/scripts/start/validateSession.ts
@@ -1,8 +1,8 @@
-import { FileSystem, Auth, Models } from "@slicemachine/core";
+import { Auth, Models, NodeUtils } from "@slicemachine/core";
 import preferWroomBase from "@lib/utils/preferWroomBase";
 
 export function validateSession(cwd: string): Promise<Models.UserInfo | null> {
-  const manifest = FileSystem.retrieveManifest(cwd);
+  const manifest = NodeUtils.retrieveManifest(cwd);
 
   // should not happen, manifest has been validated before.
   if (!manifest.exists || !manifest.content) return Promise.resolve(null);
